Refactor of FilterSideBar
