+++ web-page/src/components/FilterSideBar/FiltersApplied/index.js
@@ -4,11 +4,11 @@ import { IDENTIFIER_PREFIX } from '../FilterButtonsRow'
 
 
 const FiltersApplied = (props) => {
-    const { filters, onFilterRemove } = props
+    const { filters, onFilterRemoved } = props
     const content = filters
     .filter( ({ id }) => !id.startsWith(IDENTIFIER_PREFIX))
     .map(
-        ({ id }) => <FilterChip key={id} text={id} onRemove={onFilterRemove} />
+        ({ id }) => <FilterChip key={id} text={id} onRemove={() => onFilterRemoved(id)} />
     )
 
     return (
